Move the Callers hook to the test file that uses it
